+++ main/HSSF/Model/PolygonShape.cs
@@ -135,12 +135,12 @@ namespace NPOI.HSSF.Model
 
             ObjRecord obj = new ObjRecord();
             CommonObjectDataSubRecord c = new CommonObjectDataSubRecord();
-            c.ObjectType=CommonObjectType.MICROSOFT_OFFICE_DRAWING;
-            c.ObjectId=((short)(shapeId));
-            c.IsLocked=(true);
-            c.IsPrintable=(true);
-            c.IsAutoFill=(true);
-            c.IsAutoline=(true);
+            c.ObjectType = CommonObjectType.MICROSOFT_OFFICE_DRAWING;
+            c.ObjectId = shapeId;
+            c.IsLocked = true;
+            c.IsPrintable = true;
+            c.IsAutoFill = true;
+            c.IsAutoline = true;
             EndSubRecord e = new EndSubRecord();
 
             obj.AddSubRecord(c);
